🦺 Add param filter to submission route
